double to decimal in money calculations

+++ LykkePay/Models/AssetsPaiRatesResponseModel.cs
@@ -7,8 +7,8 @@ namespace LykkePay.Models
     public class AssetsPaiRatesResponseModel
     {
             public string assetPair { get; set; }
-            public float bid { get; set; }
-            public float ask { get; set; }
+            public decimal bid { get; set; }
+            public decimal ask { get; set; }
             public int accuracy { get; set; }
     }
 }
